refactor(sdk)!: move Api/Http/Auth types to iii-sdk/types (Node)
BREAKING CHANGE: ApiRequest, ApiResponse, HttpRequest, HttpResponse, AuthInput and
AuthResult are no longer exported from the 'iii-sdk' root. Import them from 'iii-sdk/types'.
